requirements.txt
# 核心依赖
feedparser==6.0.11
openai==1.14.0
python-dotenv==1.0.1
httpx==0.27.0
pyyaml==6.0.1

# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.30.0

# 数据库
sqlalchemy==2.0.35
